Bug 1496359 - Don't set _GLIBCXX_DEBUG in webrtc.org build config; r=ng
This define is not set in the current gn generated json and moz.build files.
If I rerun gn, this define ends up being set, and the build will fail with
a variety of link time errors.

My guess is that enable_iterator_debugging was not set when I last ran gn for
x64 Linux, and that it was subsequently enabled without regenerating the gn
files and noticing that it causes problems.

Differential Revision: https://phabricator.services.mozilla.com/D27367
Mercurial Revision: https://hg.mozilla.org/mozilla-central/rev/ceed4e0b33737af65853e8cff1242b8b74060b41